@@ -106,7 +106,7 @@ vm_size_t FIRCLSHostGetPageSize(void) {
 
 // This comes from the Apple documentation here:
 // https://developer.apple.com/documentation/apple_silicon/about_the_rosetta_translation_environment
-bool FIRCLSHostIsRosettaTranslated() {
+bool FIRCLSHostIsRosettaTranslated(void) {
 #if TARGET_OS_MAC
   int result = 0;
   size_t size = sizeof(result);
