@@ -16,7 +16,6 @@ from __future__ import division
 
 import math
 import multiprocessing.pool
-import six
 import sys
 import threading
 
@@ -55,7 +54,9 @@ class Result(object):
 
   def __init__(self, num_errors, output):
     self.errors = num_errors
-    self.output = six.ensure_text(output)
+    self.output = (output
+                   if isinstance(output, str)
+                   else output.decode('utf8', errors='replace'))
 
   @staticmethod
   def from_list(errors):
