@@ -22,8 +22,6 @@ import re
 import subprocess
 import string
 
-import six
-
 NO_HEADING = 'PRODUCT HAS NO HEADING'
 
 
@@ -78,8 +76,8 @@ def main():
 
 
 def find_local_repo():
-  url = six.ensure_text(
-      subprocess.check_output(['git', 'config', '--get', 'remote.origin.url']))
+  url = subprocess.check_output(['git', 'config', '--get', 'remote.origin.url'],
+                                text=True, errors='replace')
 
   # ssh or https style URL
   m = re.match(r'^(?:git@github\.com:|https://github\.com/)(.*)\.git$', url)
